pyproject.toml
[tool.poetry]
name = "oz-3"
version = "0.1.0"
description = ""
authors = ["seungwon0113 <dhkd0113@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-restframework = "^0.0.1"
isort = "^5.13.2"
mypy = "^1.13.0"
psycopg2-binary = "^2.9.10"
django-stubs = "^5.1.1"
transaction = "^5.0"
djangorestframework-stubs = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.base"

[tool.poetry.group.dev.dependencies]
drf-spectacular = "^0.27.2"
black = "^24.10.0"
coverage = "^7.6.4"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
